feat: make two separate apis for querying
