+++ src/commands/client/query/query-many/index.ts
@@ -1,0 +1,22 @@
+import { Command } from "@commands/command";
+import {
+  QueryCommandOptions,
+  QueryCommandPayload,
+  QueryEndpointVariants,
+  QueryResult,
+} from "../types";
+
+export class QueryManyCommand<TMetadata> extends Command<QueryResult<TMetadata>[][]> {
+  constructor(payload: QueryCommandPayload[], options?: QueryCommandOptions) {
+    let endpoint: QueryEndpointVariants = "query";
+
+    const hasData = payload.some((p) => p.data);
+    endpoint = hasData ? "query-data" : "query";
+
+    if (options?.namespace) {
+      endpoint = `${endpoint}/${options.namespace}`;
+    }
+
+    super(payload, endpoint);
+  }
+}
