+++ src/commands/types.ts
@@ -1,5 +1,5 @@
 export { type Vector } from "./client/types";
 export { type RangeResult } from "./client/range/index";
 export { type FetchResult } from "./client/fetch/index";
-export { type QueryResult } from "./client/query/index";
+export { type QueryResult } from "./client/query/types";
 export { type InfoResult } from "./client/info/index";
